requirements.txt
streamlit
pylint
pytest
click
black
python-dotenv
openai